2983: Use result types for Brush::moveBoundary
This commit changes Brush::moveBoundary to be a const member, and to
return a new brush instead of modifying this. We also remove
Brush::canMoveBoundary because it is redundant in most use cases now.
Finally Brush::moveBoundary now returns either a brush or an error
indicating why the operation failed.